delete some files
